ep-24

+++ app/Http/Controllers/JobController.php
@@ -1,7 +1,12 @@
 <?php
 
     namespace App\Http\Controllers;
+    use App\Models\User;
     use App\Models\Job;
+    use Illuminate\Support\Facades\Auth;
+    use Illuminate\Support\Facades\Gate;
+    use Illuminate\Support\Facades\Mail;
+    use App\Mail\JobPosted;
 
     class JobController extends Controller
     {
@@ -31,11 +36,16 @@
                 'salary' => ['required',],
             ]);
 
-            Job::create([
+            $job = Job::create([
                 'title' => request('title'),
                 'salary' => request('salary'),
                 'employer_id' => 1
             ]);
+
+            Mail::to($job->employer->user)->send(
+                new JobPosted($job)
+            );
+
 
             return redirect('/jobs');
 
@@ -43,12 +53,16 @@
 
         public function edit(Job $job)
         {
+
+            Gate::authorize('edit-job', $job);
+
             return view('jobs.edit', ['job' => $job]);
+
         }
 
         public function update(Job $job)
         {
-            //Authorise the request (on hold...)
+            Gate::authorize('edit-job', $job);
 
             request()->validate([
                 'title' => ['required', 'min:3'],
@@ -66,7 +80,7 @@
 
         public function destroy(Job $job)
         {
-            //Authorise the request (on hold...)
+            Gate::authorize('edit-job', $job);
 
             $job->delete();
